moved interpolation logic into separate fie
