+++ packages/cli-lib/lib/lang.js
@@ -3,19 +3,9 @@ const path = require('path');
 const fs = require('fs-extra');
 const yaml = require('js-yaml');
 const { logger } = require('../logger');
-// const { loadHandlebarsCustomHelpers } = require('./handlebarsCustomHelpers');
+const { interpolate } = require('./interpolation');
 
 const MISSING_LANGUAGE_DATA_PREFIX = '[Missing language data]';
-const DELIMITERS = {
-  interpolate: {
-    start: '{{',
-    end: '}}',
-  },
-  helpers: {
-    start: '#',
-    end: '/',
-  },
-};
 
 let locale;
 let languageObj;
@@ -75,54 +65,6 @@ const getTextValue = lookupDotNotation => {
   }
 
   return textValue;
-};
-
-const isHelperIdentifier = identifier => {
-  return (
-    identifier.startsWith(DELIMITERS.helpers.start) ||
-    identifier.startsWith(DELIMITERS.helpers.end)
-  );
-};
-
-const interpolate = (stringValue, interpolationData) => {
-  console.log('interpolationData: ', interpolationData);
-  const interpolationIdentifierRegEx = new RegExp(
-    `${DELIMITERS.interpolate.start}(.*?)${DELIMITERS.interpolate.end}`,
-    'g'
-  );
-  const replaceQueue = [];
-  let match;
-
-  while ((match = interpolationIdentifierRegEx.exec(stringValue)) != null) {
-    const { 0: matchedText, 1: rawIdentifier, index } = match;
-    const identifier = rawIdentifier.trim();
-
-    if (identifier && !isHelperIdentifier(identifier)) {
-      console.log({
-        identifier,
-        matchedText,
-        index,
-        helper: isHelperIdentifier(identifier),
-        replaceWith: interpolationData[identifier],
-      });
-      replaceQueue.unshift(theString => {
-        console.log('theString: ', theString);
-        const newString = `${theString.slice(0, index)}${interpolationData[
-          identifier
-        ] || ''}${theString.slice(index + matchedText.length)}`;
-        console.log('newString: ', newString);
-        return newString;
-      });
-    }
-  }
-
-  const compiledString = replaceQueue.reduce(
-    (currentValue, replaceFn) => replaceFn(currentValue),
-    stringValue
-  );
-  console.log('compiledString: ', compiledString);
-
-  return compiledString;
 };
 
 const getInterpolatedValue = (textValue, interpolationData) => {
